refactor: tidy api routes

+++ witwiki_server/src/models/recent_tags.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 
-#[derive(Debug, Serialize)]
+#[derive(Debug, Serialize, sqlx::Type)]
 pub struct RecentTag {
-    pub id: usize,
+    pub id: i64,
     pub tag: String,
 }
