+++ witwiki_server/src/post.rs
@@ -2,12 +2,12 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize)]
 pub struct Post {
-    pub id: usize,
-    pub user_id: usize,
+    pub id: i64,
+    pub user_id: i64,
     pub body: String,
     pub title: String,
-    pub created_at: usize,
-    pub updated_at: usize,
+    pub created_at: i64,
+    pub updated_at: i64,
     pub slug: String,
 }
